page routing set up, second rendering also fixed from main branch
